refactor: Songモデルから検索とカテゴリ機能を関心事として分離
- AlgoliaSearchable: Algolia検索に関する設定とメソッドを抽出
- Categorizable: 原曲カテゴリ分類に関する定数とメソッドを抽出
- Songモデルの責務を明確化し、保守性を向上
