+++ app/models/song.rb
@@ -1,5 +1,6 @@
 class Song < ApplicationRecord
-  include AlgoliaSearch
+  include AlgoliaSearchable
+  include Categorizable
   include ParallelProcessor
 
   has_one :song_with_dam_ouchikaraoke, dependent: :destroy
@@ -30,122 +31,15 @@ class Song < ApplicationRecord
     "https://www.joysound.com/web/search/song/225456", # Crazy speed Hight 作曲:龍5150
     "https://www.joysound.com/web/search/song/225449"  # 愛き夜道 feat. ランコ(豚乙女)、雨天決行／魂音泉 作曲:U2，Coro
   ].freeze
-  ORIGINAL_TYPE = {
-    windows: "01. Windows作品",
-    pc98: "02. PC-98作品",
-    zuns_music_collection: "03. ZUN's Music Collection",
-    akyus_untouched_score: "04. 幺樂団の歴史　～ Akyu's Untouched Score",
-    commercial_books: "05. 商業書籍",
-    other: "06. その他"
-  }.freeze
 
   def self.ransackable_attributes(_auth_object = nil)
     ["title"]
-  end
-
-  algoliasearch index_name: ENV.fetch('ALGOLIA_INDEX_NAME', nil), unless: :deleted? do
-    attribute :title
-    attribute :reading_title do
-      title_reading || ''
-    end
-    attribute :display_artist do
-      {
-        name: display_artist.name,
-        reading_name: display_artist.name_reading,
-        reading_name_hiragana: display_artist.name_reading.tr('ァ-ン', 'ぁ-ん'),
-        karaoke_type: display_artist.karaoke_type,
-        url: display_artist.url
-      }
-    end
-    attribute :original_songs do
-      original_songs_json(original_songs)
-    end
-    attribute :karaoke_type
-    attribute :karaoke_delivery_models do
-      karaoke_delivery_models_json
-    end
-    attribute :circle do
-      {
-        name: display_artist.circles.first&.name || ''
-      }
-    end
-    attribute :url
-    attribute :song_number do
-      song_number.presence
-    end
-    attribute :delivery_deadline_date do
-      song_with_joysound_utasuki&.delivery_deadline_date&.strftime("%Y/%m/%d")
-    end
-    attribute :musicpost_url do
-      song_with_joysound_utasuki&.url
-    end
-    attribute :ouchikaraoke_url do
-      song_with_dam_ouchikaraoke&.url
-    end
-    attribute :videos
   end
 
   def touhou?
     return false if original_songs.blank?
 
     original_songs.all? { _1.title != 'オリジナル' } && !original_songs.all? { _1.title == 'その他' }
-  end
-
-  def first_category(original)
-    ORIGINAL_TYPE[original.original_type.to_sym]
-  end
-
-  def second_category(original)
-    "#{first_category(original)} > #{format('%#04.1f', original.series_order)}. #{original.short_title}"
-  end
-
-  def third_category(original_song)
-    original = original_song.original
-    "#{second_category(original)} > #{format('%02d', original_song.track_number)}. #{original_song.title}"
-  end
-
-  def original_songs_json(original_songs)
-    original_songs.map do |os|
-      {
-        title: os.title,
-        original: {
-          title: os.original.title,
-          short_title: os.original.short_title
-        },
-        'categories.lvl0': first_category(os.original),
-        'categories.lvl1': second_category(os.original),
-        'categories.lvl2': third_category(os)
-      }
-    end
-  end
-
-  def karaoke_delivery_models_json
-    karaoke_delivery_models.map do |kdm|
-      {
-        name: kdm.name,
-        karaoke_type: kdm.karaoke_type
-      }
-    end
-  end
-
-  def videos
-    v = []
-    if youtube_url.present?
-      m = /(?<=\?v=)(?<id>[\w\-_]+)(?!=&)/.match(youtube_url)
-      v.push({ type: "YouTube", url: youtube_url, id: m[:id] })
-    end
-    if nicovideo_url.present?
-      m = %r{(?<=watch/)(?<id>[s|n]m\d+)(?!=&)}.match(nicovideo_url)
-      v.push({ type: "ニコニコ動画", url: nicovideo_url, id: m[:id] })
-    end
-    v
-  end
-
-  def deleted?
-    return true if original_songs.blank?
-
-    original_song_titles = original_songs.map(&:title)
-    original_song_titles.include?("オリジナル") || original_song_titles.include?("その他")
   end
 
   def self.not_set_original_song
